csharp-core: edited homecontroller
