+++ web/documentserver-example/csharp-core/OnlineEditorsExampleNetCore/Startup.cs
@@ -38,6 +38,10 @@ namespace OnlineEditorsExampleNetCore
 
             services.AddHttpContextAccessor();
 
+            services.AddDistributedMemoryCache();
+
+            services.AddSession();
+
             services.AddControllersWithViews();
 
             // If using Kestrel:
@@ -56,6 +60,8 @@ namespace OnlineEditorsExampleNetCore
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
         public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
+            app.UseSession();
+
             if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
@@ -65,7 +71,7 @@ namespace OnlineEditorsExampleNetCore
                 app.UseExceptionHandler("/Home/Error");
             }
 
-            app.UseMyMiddleware();
+            //app.UseMyMiddleware();
 
             app.UseStaticFiles();
