+++ web/documentserver-example/csharp-core/OnlineEditorsExampleNetCore/Views/Home/Index.cshtml
@@ -276,7 +276,7 @@
                                                             </a>
                                                         </td>
                                                         <td class="contentCells contentCells-icon contentCells-shift">
-                                                            <a href="?type=download&fileName=@HttpUtility.UrlEncode(storedFile.Name)">
+                                                            <a href="download&fileName=@HttpUtility.UrlEncode(storedFile.Name)">
                                                                 <img class="icon-download" src="~/img/download-24.png" alt="Download" title="Download" />
                                                             </a>
                                                         </td>
